% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_ferroptosis.R
\name{barploths}
\alias{barploths}
\title{Barplot ferroptosis style ferroviz (Up/Down proportions)}
\usage{
barploths(resulths, ferrdb, fc = 0.1, size = 16)
}
\arguments{
\item{resulths}{Data frame limma contenant logFC, adj.P.Val.}

\item{fc}{Seuil absolu de logFC (défaut = 0.10).}

\item{size}{Taille du texte du graphique (défaut = 16).}

\item{ferrdbhs}{Data frame FerrDb contenant au minimum "hs.gene" et "class".}
}
\value{
Un objet ggplot2 représentant le barplot.
}
\description{
Cette fonction filtre les gènes significatifs selon un seuil de logFC et de
p-value ajustée, annote avec FerrDb, classe les gènes en Up/Down, puis génère
un barplot proportionnel (position = "fill") montrant la proportion de gènes
ferroptosis (driver/suppressor/both) dans chaque direction de régulation.
}
\examples{
barploths(resulths, ferrdb, fc = 0.25, size = 16)

}
