% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ferroptosis_table.R
\name{export_ferroptosis_table}
\alias{export_ferroptosis_table}
\title{Export table of significant ferroptosis-related genes (ferroviz style)}
\usage{
export_ferroptosis_table(resulths, ferrdb, fc = 0.1)
}
\arguments{
\item{resulths}{Data frame limma contenant logFC, adj.P.Val.}

\item{ferrdb}{Data frame FerrDb contenant au minimum "Symbol" et "cat".}

\item{fc}{Seuil absolu de logFC (défaut = 0.10).}
}
\value{
Un data.frame contenant les gènes ferroptosis significatifs.
}
\description{
Cette fonction filtre les gènes significatifs selon un seuil de logFC et de
p-value ajustée, annote avec FerrDb, classe les gènes en Up/Down, puis
retourne un data.frame contenant uniquement les gènes ferroptosis
(driver / suppressor / both).
}
\examples{
tbl <- export_ferroptosis_table(resulths, ferrdb, fc = 0.1)
head(tbl)

}
