% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_ferroptosis.R
\name{volcano_ferroptosis}
\alias{volcano_ferroptosis}
\title{Volcano plot annoté avec FerrDb (driver / suppressor / both)}
\usage{
volcano_ferroptosis(
  resulths,
  ferrdb,
  fc_thresh = 0.5,
  p_thresh = 0.05,
  plot_title = "DEGs and ferroptosis",
  base_size = 14,
  label_size = 4
)
}
\arguments{
\item{resulths}{Data frame issu de limma contenant logFC, P.Value, adj.P.Val.}

\item{ferrdb}{Data frame FerrDb contenant au minimum la colonne "Symbol" et "cat".}

\item{fc_thresh}{Seuil absolu de logFC (défaut = 0.5).}

\item{p_thresh}{Seuil de p-value ajustée (défaut = 0.05).}

\item{plot_title}{Titre du graphique (défaut = "DEGs and ferroptosis").}

\item{base_size}{Taille de base du thème ggplot2 (défaut = 14).}

\item{label_size}{Taille des labels ggrepel (défaut = 4).}
}
\value{
Un objet ggplot2 représentant le volcano plot.
}
\description{
Cette fonction génère un volcano plot à partir d'un objet limma (resulths)
et d'une annotation FerrDb (driver, suppressor, both).
}
\examples{
load("ferrdb.rda")
volcano_ferroptosis(resulths, ferrdb, plot_title = "Mon Volcano",fc_thresh=0.25,base_size=16)

}
